refactor: remove root builder shims
