--- tensorrt_model_connect/tensorrt_model_connect/causal_vae_3d_builder.py
@@ -1,7 +1,0 @@
-"""Compatibility shim for the wan_t2v family-owned causal_vae_3d_builder implementation."""
-
-import sys as _sys
-
-from .families.wan_t2v import causal_vae_3d_builder as _impl
-
-_sys.modules[__name__] = _impl
