--- tensorrt_model_connect/tensorrt_model_connect/convbert_encoder_builder.py
@@ -1,7 +1,0 @@
-"""Compatibility shim for the ConvBERT family-owned builder."""
-
-import sys as _sys
-
-from .families.convbert import builder as _impl
-
-_sys.modules[__name__] = _impl
