--- tensorrt_model_connect/tensorrt_model_connect/dual_profile_decoder_builder.py
@@ -1,7 +1,0 @@
-"""Compatibility shim for the llama family-owned dual_profile_decoder_builder implementation."""
-
-import sys as _sys
-
-from .families.llama import dual_profile_decoder_builder as _impl
-
-_sys.modules[__name__] = _impl
