--- tensorrt_model_connect/tensorrt_model_connect/encodec_builder.py
@@ -1,7 +1,0 @@
-"""Compatibility shim for the bark family-owned encodec_builder implementation."""
-
-import sys as _sys
-
-from .families.bark import encodec_builder as _impl
-
-_sys.modules[__name__] = _impl
