--- tensorrt_model_connect/tensorrt_model_connect/encoder_builder.py
@@ -1,7 +1,0 @@
-"""Compatibility shim for the bert family-owned encoder_builder implementation."""
-
-import sys as _sys
-
-from .families.bert import encoder_builder as _impl
-
-_sys.modules[__name__] = _impl
