--- tensorrt_model_connect/tensorrt_model_connect/flux_vae_builder.py
@@ -1,7 +1,0 @@
-"""Compatibility shim for the flux family-owned flux_vae_builder implementation."""
-
-import sys as _sys
-
-from .families.flux import flux_vae_builder as _impl
-
-_sys.modules[__name__] = _impl
