--- tensorrt_model_connect/tensorrt_model_connect/standard_dit_builder.py
@@ -1,7 +1,0 @@
-"""Compatibility shim for the pixart family-owned standard_dit_builder implementation."""
-
-import sys as _sys
-
-from .families.pixart import standard_dit_builder as _impl
-
-_sys.modules[__name__] = _impl
