--- tensorrt_model_connect/tensorrt_model_connect/vae_2d_builder.py
@@ -1,7 +1,0 @@
-"""Compatibility shim for the pixart family-owned vae_2d_builder implementation."""
-
-import sys as _sys
-
-from .families.pixart import vae_2d_builder as _impl
-
-_sys.modules[__name__] = _impl
