--- tensorrt_model_connect/tensorrt_model_connect/vision_encoder_builder.py
@@ -1,7 +1,0 @@
-"""Compatibility shim for the Qwen-VL family-owned vision encoder builder."""
-
-import sys as _sys
-
-from . import qwen_vl_vision_builder as _impl
-
-_sys.modules[__name__] = _impl
