--- tensorrt_model_connect/tensorrt_model_connect/xlnet_builder.py
@@ -1,7 +1,0 @@
-"""Compatibility shim for the xlnet family-owned xlnet_builder implementation."""
-
-import sys as _sys
-
-from .families.xlnet import xlnet_builder as _impl
-
-_sys.modules[__name__] = _impl
